Rename _metadata_import.py -> _metadata_import_utils.py
(as per Chuck's suggestion)
